use interface_ip helper method instead of shelling out
