+++ agent/lib/kontena/helpers/node_helper.rb
@@ -1,6 +1,3 @@
-require 'socket'
-require 'ipaddr'
-
 module Kontena
   module Helpers
     module NodeHelper
@@ -15,18 +12,6 @@ module Kontena
         if response.status == 200
           return JSON.load(response.body)
         end
-        nil
-      end
-
-      # @param [String] iface
-      # @return [String, NilClass]
-      def interface_ip(iface)
-        sock = UDPSocket.new
-        buf = [iface,""].pack('a16h16')
-        sock.ioctl(SIOCGIFADDR, buf);
-        sock.close
-        buf[20..24].unpack("CCCC").join(".")
-      rescue Errno::ENODEV
         nil
       end
